Suffix with -main
