+++ build_tools/embed_data/BUILD
@@ -24,7 +24,7 @@ package(
 
 cc_binary(
     name = "generate_cc_embed_data",
-    srcs = ["generate_cc_embed_data.cc"],
+    srcs = ["generate_cc_embed_data_main.cc"],
     deps = [
         "@com_google_absl//absl/flags:flag",
         "@com_google_absl//absl/flags:parse",
