+++ build_tools/embed_data/generate_cc_embed_data_main.cc
