+++ iree/tools/CMakeLists.txt
@@ -51,7 +51,7 @@ iree_cc_binary(
   OUT
     iree-benchmark-module
   SRCS
-    "iree-benchmark-module_main.cc"
+    "iree-benchmark-module-main.cc"
   DEPS
     ::vm_util
     absl::flags
@@ -74,7 +74,7 @@ iree_cc_binary(
   OUT
     iree-dump-module
   SRCS
-    "iree-dump-module_main.cc"
+    "iree-dump-module-main.cc"
   DEPS
     flatbuffers
     iree::base::file_io
@@ -89,7 +89,7 @@ iree_cc_binary(
   OUT
     iree-run-module
   SRCS
-    "iree-run-module_main.cc"
+    "iree-run-module-main.cc"
   DEPS
     ::vm_util
     absl::flags
@@ -232,7 +232,7 @@ if(${IREE_BUILD_COMPILER})
     NAME
       iree_opt_main
     SRCS
-      "iree-opt_main.cc"
+      "iree-opt-main.cc"
     DEPS
       ::init_compiler_modules
       ::init_iree_passes_and_dialects
@@ -278,7 +278,7 @@ if(${IREE_BUILD_COMPILER})
     NAME
       iree_translate_main
     SRCS
-      "iree-translate_main.cc"
+      "iree-translate-main.cc"
     DEPS
       ::init_compiler_modules
       ::init_iree_passes_and_dialects
@@ -316,7 +316,7 @@ if(${IREE_BUILD_COMPILER})
     OUT
       iree-run-mlir
     SRCS
-      "iree-run-mlir_main.cc"
+      "iree-run-mlir-main.cc"
     DEPS
       ::init_compiler_modules
       ::init_iree_passes_and_dialects
